+++ index.html
@@ -27,7 +27,7 @@
 
   </div>
 
-  <script src="./main.js"></script>
+  <script type="module" src="./main.js"></script>
   
 </body>
 </html>
